% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateByfileBIN.R
\name{aggregateByfileBIN}
\alias{aggregateByfileBIN}
\title{Title}
\usage{
aggregateByfileBIN(dfDiario, aggregateCalendar2BIN, partition = NULL)
}
\arguments{
\item{dfDiario}{}

\item{aggregateCalendar2BIN}{}

\item{partition}{}
}
\description{
Title
}
