% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicSummary.R
\name{basicSummary}
\alias{basicSummary}
\title{Title}
\usage{
basicSummary(listBouts, listCalendarsAndBouts, partition = NULL,
  label = "partition")
}
\arguments{
\item{listBouts}{}

\item{listCalendarsAndBouts}{}

\item{partition}{}

\item{label}{}
}
\description{
Title
}
