% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicSummaryOfBouts.R
\name{basicSummaryOfBouts}
\alias{basicSummaryOfBouts}
\title{Title}
\usage{
basicSummaryOfBouts(listBouts, listCalendarsAndBouts)
}
\arguments{
\item{listBouts}{}

\item{listCalendarsAndBouts}{}
}
\description{
Title
}
