% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicSummaryOfPartitions.R
\name{basicSummaryOfPartitions}
\alias{basicSummaryOfPartitions}
\title{Title}
\usage{
basicSummaryOfPartitions(listBouts, listCalendarsAndBouts, partition,
  label = "partition")
}
\arguments{
\item{listBouts}{}

\item{listCalendarsAndBouts}{}

\item{partition}{}

\item{label}{}
}
\description{
Title
}
