% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicVariables.R
\name{bedGGIR}
\alias{bedGGIR}
\title{summarises acc_onset + acc_wake de fase 5 ggir.
Criteria: GGIR. Viene de los fase 5. FALTA describirlo.}
\usage{
bedGGIR(df)
}
\arguments{
\item{df}{Dataframe of epochs with one columns: datetime}
}
\value{
a list with a summary of the periods matching the criteria
}
\description{
summarises acc_onset + acc_wake de fase 5 ggir.
Criteria: GGIR. Viene de los fase 5. FALTA describirlo.
}
