% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeCalendars.R
\name{computeCalendars}
\alias{computeCalendars}
\title{Dado una lista de bouts, y una lista que contiene funciones de generación de calendarios,
devuelve lista de calendarios}
\usage{
computeCalendars(bouts, listFunctionalCalendars, mc.cores = numCores(4))
}
\arguments{
\item{bouts}{}

\item{listFunctionalCalendars}{}

\item{mc.cores}{}
}
\description{
Dado una lista de bouts, y una lista que contiene funciones de generación de calendarios,
devuelve lista de calendarios
}
