% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeInvalidCalendars.R
\name{computeInvalidCalendars}
\alias{computeInvalidCalendars}
\title{Calcula un dataframe con los calendarios que no son válidos. Depende de una lista de bouts de entrada y de calendarios. 
Si alguno hace el papel de contener la información de cual es la cama del día, hay que indicar su nombre}
\usage{
computeInvalidCalendars(calendars, bouts, labelbed = "bed")
}
\arguments{
\item{calendars}{}

\item{bouts}{}

\item{labelbed}{}
}
\description{
Calcula un dataframe con los calendarios que no son válidos. Depende de una lista de bouts de entrada y de calendarios. 
Si alguno hace el papel de contener la información de cual es la cama del día, hay que indicar su nombre
}
