% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeValidCalendars.R
\name{computeValidCalendars}
\alias{computeValidCalendars}
\title{Dado una lista de bouts, y una lista que contiene funciones de generación de calendarios,
devuelve lista de calendarios}
\usage{
computeValidCalendars(bouts, listFunctionalCalendars, labelbed = "bed",
  progreso = NULL)
}
\arguments{
\item{bouts}{}

\item{listFunctionalCalendars}{}

\item{labelbed}{}

\item{progreso}{}
}
\description{
Dado una lista de bouts, y una lista que contiene funciones de generación de calendarios,
devuelve lista de calendarios
}
