% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectOverDistance.R
\name{connectOverDistance}
\alias{connectOverDistance}
\title{Connect intervals that are close enough
Generates a smaller dataframe of connected intervals using a data frame of given intervals and a distance that allows tow of them to be connected}
\usage{
connectOverDistance(interval, distance = dminutes(30))
}
\arguments{
\item{interval}{Initial dataframe of intervals to be connected}

\item{distance}{distance between ttwo intervals to allo the connection of both in just one.}
}
\value{
A dataframe of intervals, having less or equal rows tan the original
}
\description{
Connect intervals that are close enough
Generates a smaller dataframe of connected intervals using a data frame of given intervals and a distance that allows tow of them to be connected
}
