% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectOverInterval.R
\name{connectOverInterval}
\alias{connectOverInterval}
\title{Connect intervals of a first dataframe using a second dataframe of intervals}
\usage{
connectOverInterval(interval1, interval2)
}
\arguments{
\item{interval1}{first dataframe}

\item{interval2}{second dataframe}
}
\value{
A dataframe of intervals representing the connected intervals
}
\description{
Connect the intervals of a first dataframe given that the can be considered connected if the separation between two of them are covered by a interval of a second dataframe.
}
