% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectOverVariableDistance.R
\name{connectOverVariableDistance}
\alias{connectOverVariableDistance}
\title{Connect intervals that are close enough
Generates a smaller dataframe of connected intervals using a data frame of given intervals and a distance that allows tow of them to be connected}
\usage{
connectOverVariableDistance(interval, fraccion = 0.33)
}
\arguments{
\item{interval}{Initial dataframe of intervals to be connected}

\item{distance}{distance between ttwo intervals to allo the connection of both in just one.}
}
\value{
A dataframe of intervals, having less or equal rows tan the original
}
\description{
Connect intervals that are close enough
Generates a smaller dataframe of connected intervals using a data frame of given intervals and a distance that allows tow of them to be connected
}
