% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterioBout.R
\name{criterioBout}
\alias{criterioBout}
\title{Indicates which rows meets the criteria for be considered part of a Bout meeting certain criteria}
\usage{
criterioBout(df, pctBouts = 1, durBoutMin = dseconds(5),
  durEpoch = dseconds(5))
}
\arguments{
\item{df}{data frame with columns .criterio and eventually .criterioNW (that represents NonWear time as TRUE/FALSE)}

\item{durBoutMin}{minimum amount of time that the condition muest be met to be aconsidered a Bout}

\item{durEpoch}{amount of time that represents each row of the dataframe (duration of a epoch usually)}

\item{pctBout}{represents fraction of time that the .criterio must be TRUE}
}
\value{
a boolean vector (TRUE/FAlSE) indicating if the condition of belonging to a Bout is met.
}
\description{
Generates a vector that indicates for each row of a dataframe (usually epoch or BIN file)
if that row meets the criteria to be considered part of a Bout (veryfy a criteria for ca percent of time of a minimum duration),
}
