% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterioENMO.R
\name{criterioENMO}
\alias{criterioENMO}
\title{Indicates which rows meets certain limits for ENMO}
\usage{
criterioENMO(df, limInf = 0, limSup = Inf, useNW = TRUE)
}
\arguments{
\item{df}{data frame with columns ENMO and eventually .criterioNW (that represents NonWear time as TRUE/FALSE)}

\item{limInf}{inferior limit for ENMO}

\item{limSup}{superior limit for ENMO}
}
\value{
a boolean vector (TRUE/FAlSE) indicating if the condition is met.
}
\description{
Generates a vector that indicates for each row of a dataframe (usually epoch or BIN file) if that row verifies the condition
   limInf<=df[["ENMO"]] <=limSup
}
