% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterioNoMVPA.R
\name{criterioNoMVPA}
\alias{criterioNoMVPA}
\title{Indicates which rows can not meet the criteria for be considered part of a MVPA or higher, because there is not enough angle z mouvement (bic accelerations without change in angle,
 means probably the subject is in a car)}
\usage{
criterioNoMVPA(df, critAnglez = 5, durBoutMin = dseconds(120),
  durEpoch = dseconds(5))
}
\arguments{
\item{df}{data frame with columns ANGLEZ, ENMO}

\item{critAnglez}{represents minimum of deviation (in both directios) of angle Z that must happen in a MVPA bout}

\item{durBoutMin}{minimum amount of time that the conditions must be met.}

\item{durEpoch}{amount of time that represents each row of the dataframe (duration of a epoch usually)}
}
\value{
a boolean vector (TRUE/FAlSE) indicating if the condition of belonging to a SIB is met.
}
\description{
Generates a vector that indicates for each row of a dataframe (usually epoch or BIN file) if that row verifies the condition to be considered a SIB,
#'
}
