% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterioSIB.R
\name{criterioSIB}
\alias{criterioSIB}
\title{Indicates which rows meets the criteria for be considered part of a SIB (Sustained Inactivity Bouts)}
\usage{
criterioSIB(df, critAnglez = 5, limSup = 30/1000,
  durBoutMin = dminutes(5), durEpoch = dseconds(5))
}
\arguments{
\item{df}{data frame with columns ANGLEZ, ENMO and eventually .criterioNW (that represents NonWear time as TRUE/FALSE)}

\item{critAnglez}{represents maximum of deviation (in both directios) of angle Z that it is allowed during a SIB}

\item{limSup}{superior limit for ENMO}

\item{durBoutMin}{minimum amount of time that the conditions must be met to beconsidered a SIB period}

\item{durEpoch}{amount of time that represents each row of the dataframe (duration of a epoch usually)}
}
\value{
a boolean vector (TRUE/FAlSE) indicating if the condition of belonging to a SIB is met.
}
\description{
Generates a vector that indicates for each row of a dataframe (usually epoch or BIN file) if that row verifies the condition to be considered a SIB,
i.e Anglez variying a few degrees for a certain amount of time (durBoutMin, and ENMO lower than a low value)
}
