% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfDecorate.R
\name{dfDecorate}
\alias{dfDecorate}
\title{Take a bout or a list of them and return a list with the same structure, but adding columns
with extra information (files and variable measured) in each dataframe containing bout information.}
\usage{
dfDecorate(element, label = c(), value = c())
}
\arguments{
\item{element}{}

\item{label}{}

\item{value}{}
}
\description{
Take a bout or a list of them and return a list with the same structure, but adding columns
with extra information (files and variable measured) in each dataframe containing bout information.
}
