% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doSummary.R
\name{doSummary}
\alias{doSummary}
\title{Title}
\usage{
doSummary(intervalReference, intervalBout, by = "day")
}
\arguments{
\item{intervalReference}{}

\item{intervalBout}{}

\item{by}{}
}
\description{
Title
}
