% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epoch2Bouts.R
\name{epoch2Bouts}
\alias{epoch2Bouts}
\title{Genera una lista de todos los bouts para un dataframe con los nodos y ficheros para los que tenemos los epochs ya calculados}
\usage{
epoch2Bouts(dfTareas)
}
\arguments{
\item{dfTareas}{}
}
\description{
Genera una lista de todos los bouts para un dataframe con los nodos y ficheros para los que tenemos los epochs ya calculados
}
