% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterInvalidCalendars.R
\name{filterInvalidCalendars}
\alias{filterInvalidCalendars}
\title{Dado una lista de bouts y un calendario devuelve una lsista de calendarios donde se cumplen las reglas de
 validez. Por tanto lo que obtenemos es un alista de calendarios más limitada.}
\usage{
filterInvalidCalendars(calendars, bouts, labelbed = "bed")
}
\arguments{
\item{calendars}{}

\item{bouts}{}

\item{labelbed}{}
}
\description{
Dado una lista de bouts y un calendario devuelve una lsista de calendarios donde se cumplen las reglas de
 validez. Por tanto lo que obtenemos es un alista de calendarios más limitada.
}
