% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicCalendars.R
\name{genCalendar_bed}
\alias{genCalendar_bed}
\title{Title}
\usage{
genCalendar_bed(bouts, label = "bed", durMin = dminutes(160),
  durMax = dhours(24))
}
\arguments{
\item{bouts}{}

\item{label}{}

\item{durMin}{}

\item{durMax}{}
}
\description{
Title
}
