% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicCalendars.R
\name{genCalendar_complementary}
\alias{genCalendar_complementary}
\title{Title}
\usage{
genCalendar_complementary(bouts, calendar, label = "complementary")
}
\arguments{
\item{bouts}{}

\item{calendar}{}

\item{label}{}
}
\description{
Title
}
