% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicCalendars.R
\name{genCalendar_complementaryInternal}
\alias{genCalendar_complementaryInternal}
\title{Title}
\usage{
genCalendar_complementaryInternal(bouts, calendar,
  label = "complementaryInternal")
}
\arguments{
\item{bouts}{}

\item{calendar}{}

\item{label}{}
}
\description{
Title
}
