% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicCalendars.R
\name{genCalendar_daily}
\alias{genCalendar_daily}
\title{Title}
\usage{
genCalendar_daily(bouts, label = "daily", first = TRUE, last = TRUE,
  offsetLabels = dhours(0), starts = dhours(0), duration = dhours(23)
  + dminutes(59) + dseconds(59))
}
\arguments{
\item{bouts}{}

\item{label}{}

\item{first}{}

\item{last}{}

\item{offsetLabels}{}

\item{starts}{}

\item{duration}{}
}
\description{
Title
}
