% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicCalendars.R
\name{genCalendar_HorarioToHorarioFrom}
\alias{genCalendar_HorarioToHorarioFrom}
\title{Title}
\usage{
genCalendar_HorarioToHorarioFrom(horario1, horario2,
  label = "horario_1_2")
}
\arguments{
\item{horario1}{}

\item{horario2}{}

\item{label}{}
}
\description{
Title
}
