% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterio2Summary.R, R/getSummary.R
\name{getSummary}
\alias{getSummary}
\title{Create a summery for the column .criterio of a data frame representing accelerometer data}
\usage{
getSummary(df, offset = dhours(0), minimoHorasValidas = 20,
  maximoHorasNonWear = 2)

getSummary(df, offset = dhours(0), minimoHorasValidas = 20,
  maximoHorasNonWear = 2)
}
\arguments{
\item{df}{data frame with columns timestamp and .criterio}

\item{offset}{Will add a quantity to the timestamp to assign the instant of time to the appropiate day}

\item{minimoHorasValidas}{explicar}

\item{maximoHorasNonWear}{explicar}

\item{Variable}{minimum amount of time that the condition muest be met to be aconsidered a Bout}

\item{durEpoch}{amount of time that represents each row of the dataframe (duration of a epoch usually)}

\item{df}{data frame with columns timestamp, .criterioRaw & .criterioBout}

\item{offset}{Will add a quantity to the timestamp to assign the instant of time to the appropiate day}

\item{Variable}{minimum amount of time that the condition muest be met to be aconsidered a Bout}

\item{durEpoch}{amount of time that represents each row of the dataframe (duration of a epoch usually)}

\item{minimoHorasValidas}{explicar}

\item{maximoHorasNonWear}{explicar}
}
\value{
a list with a summary. This is the info that we use to define activity variables on a daily basis and on a global value.

a list with a summary. This is the info that we use to define activity variables on a daily basis and on a global value.
}
\description{
Generates a summary for a epoch or BIN file

Generates a summary for a epoch or BIN file
}
