% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervalIntersect.R
\name{intervalIntersect}
\alias{intervalIntersect}
\title{Intersect the intervals given by two dataframes
Generates the intersection dataframe resulting from the intersection of the intervals represented by two dataframes of intervals}
\usage{
intervalIntersect(interval1, interval2)
}
\arguments{
\item{interval1}{first dataframe}

\item{interval2}{second dataframe}
}
\value{
A dataframe of intervals representing the inersection
}
\description{
Intersect the intervals given by two dataframes
Generates the intersection dataframe resulting from the intersection of the intervals represented by two dataframes of intervals
}
