% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervalMelt.R
\name{intervalMelt}
\alias{intervalMelt}
\title{Melt intervals of a first dataframe using a second dataframe of intervals}
\usage{
intervalMelt(interval1, interval2)
}
\arguments{
\item{interval1}{first dataframe}

\item{interval2}{second dataframe}
}
\value{
A dataframe of intervals representing the resulting connected intervals
}
\description{
Melt intervals of a first dataframe using a second dataframe of intervals
}
