% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervalSimplify.R
\name{intervalSimplify}
\alias{intervalSimplify}
\title{Simplify intervals of a first dataframe using a second dataframe of intervals}
\usage{
intervalSimplify(interval)
}
\arguments{
\item{interval}{first dataframe}
}
\value{
A dataframe of intervals representing the resulting connected intervals
}
\description{
Simplify intervals of a first dataframe using a second dataframe of intervals
}
