% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicVariables.R
\name{isOn}
\alias{isOn}
\title{summarises time of Non Wear.
Criteria: GGIR. Viene de los epochs, creado en fase 1. FALTA describir.}
\usage{
isOn(df)
}
\arguments{
\item{df}{Dataframe of epochs with two columns: datetime and ENMO}
}
\value{
a list with a summary of the periods matching the criteria
}
\description{
summarises time of Non Wear.
Criteria: GGIR. Viene de los epochs, creado en fase 1. FALTA describir.
}
