% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list2df.R
\name{list2df}
\alias{list2df}
\title{Convert a recursive list of dataframes to one dataframe.}
\usage{
list2df(lista, label = "")
}
\arguments{
\item{lista}{}

\item{label}{}
}
\description{
Convert a recursive list of dataframes to one dataframe.
}
