% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicVariables.R
\name{LMVPA_B5}
\alias{LMVPA_B5}
\title{Summarises bouts of at least Light Physical Activity (LMVPA, >1,5 METs).
Can be used to define inactivity BREAKS.
Criteria: 5min-bouts with >100% of at least LPA, treashold ENMO >40 mg.}
\usage{
LMVPA_B5(df)
}
\arguments{
\item{df}{Dataframe of epochs with two columns: datetime and ENMO}
}
\value{
a list with a summary of the periods matching the criteria
}
\description{
Summarises bouts of at least Light Physical Activity (LMVPA, >1,5 METs).
Can be used to define inactivity BREAKS.
Criteria: 5min-bouts with >100% of at least LPA, treashold ENMO >40 mg.
}
