% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicVariables.R
\name{MVPA_B10_H}
\alias{MVPA_B10_H}
\title{Summarises bouts of Moderate-to-Vigorous Physical Activity (MVPA, >3METs) such as brisk walking during at least 10 min.
Criteria: 10-minute-bouts with >80% of MVPA, treashold ENMO <100mg.}
\usage{
MVPA_B10_H(df)
}
\arguments{
\item{df}{Dataframe of epochs with two columns: datetime and ENMO}
}
\value{
a list with a summary of the periods matching the criteria
}
\description{
Summarises bouts of Moderate-to-Vigorous Physical Activity (MVPA, >3METs) such as brisk walking during at least 10 min.
Criteria: 10-minute-bouts with >80% of MVPA, treashold ENMO <100mg.
}
