% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicVariables.R
\name{nonWear}
\alias{nonWear}
\title{summarises Non Wear time.
Criteria: GGIR. Viene de los epochs, creado en fase 1. FALTA describirlo.}
\usage{
nonWear(df)
}
\arguments{
\item{df}{Dataframe of epochs with two columns: datetime and ENMO}
}
\value{
a list with a summary of the periods matching the criteria
}
\description{
summarises Non Wear time.
Criteria: GGIR. Viene de los epochs, creado en fase 1. FALTA describirlo.
}
