% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtenerBouts.R
\name{obtenerBouts}
\alias{obtenerBouts}
\title{Title}
\usage{
obtenerBouts(Nodo, fileBIN, desde = NA, hasta = NA,
  nucleos = numCores(4), progreso = NULL)
}
\arguments{
\item{Nodo}{}

\item{fileBIN}{}

\item{nucleos}{}
}
\description{
Title
}
