% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicVariables.R
\name{SIB2}
\alias{SIB2}
\title{Summarises Sustained Inactivity Bouts (SIB) according to the criteria by Vincent van Hees
Critera: Identify 5-min-bouts with < 5 degrees change in angle.}
\usage{
SIB2(df)
}
\arguments{
\item{df}{Dataframe of epochs with two columns: datetime and XXXXXX.}
}
\value{
a list with a summary of the periods matching the criteria.
}
\description{
Summarises Sustained Inactivity Bouts (SIB) according to the criteria by Vincent van Hees
Critera: Identify 5-min-bouts with < 5 degrees change in angle.
}
