% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryBoutsByCalendar.R
\name{summaryBoutsByCalendar}
\alias{summaryBoutsByCalendar}
\title{Dado una lista de bouts, calendarios y otra lista de que bouts queremos medir para cada calendario
nos devuelve un dataframe con un resumen de las variables de interés (bouts) medidas en sus respectivos calendarios}
\usage{
summaryBoutsByCalendar(bouts, calendars, listCalendarsAndBouts)
}
\arguments{
\item{bouts}{}

\item{calendars}{}

\item{listCalendarsAndBouts}{}
}
\description{
Dado una lista de bouts, calendarios y otra lista de que bouts queremos medir para cada calendario
nos devuelve un dataframe con un resumen de las variables de interés (bouts) medidas en sus respectivos calendarios
}
