% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryBoutsByPartition.R
\name{summaryBoutsByPartition}
\alias{summaryBoutsByPartition}
\title{Title}
\usage{
summaryBoutsByPartition(bouts, calendars, listCalendarsAndBouts, partition,
  label = "partition")
}
\arguments{
\item{bouts}{}

\item{calendars}{}

\item{listCalendarsAndBouts}{}

\item{partition}{}

\item{label}{}
}
\description{
Title
}
