% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryListOfBoutsByCalendar.R
\name{summaryListOfBoutsByCalendar}
\alias{summaryListOfBoutsByCalendar}
\title{Title}
\usage{
summaryListOfBoutsByCalendar(listBouts, listCalendarsAndBouts)
}
\arguments{
\item{listBouts}{}

\item{listCalendarsAndBouts}{}
}
\description{
Title
}
