% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryListOfBoutsByPartition.R
\name{summaryListOfBoutsByPartition}
\alias{summaryListOfBoutsByPartition}
\title{Title}
\usage{
summaryListOfBoutsByPartition(listBouts, listCalendarsAndBouts, partition,
  label = "partition")
}
\arguments{
\item{listBouts}{}

\item{listCalendarsAndBouts}{}

\item{partition}{}

\item{label}{}
}
\description{
Title
}
