% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicVariables.R
\name{VPA_B30s}
\alias{VPA_B30s}
\title{Summarises bouts of Vigorous Physical Activity (VPA, >6METs) such as jogging or running.
Criteria: 30-seconds-bouts with >80% of VPA, treashold ENMO <500mg.}
\usage{
VPA_B30s(df)
}
\arguments{
\item{df}{Dataframe of epochs with two columns: datetime and ENMO}
}
\value{
a list with a summary of the periods matching the criteria
}
\description{
Summarises bouts of Vigorous Physical Activity (VPA, >6METs) such as jogging or running.
Criteria: 30-seconds-bouts with >80% of VPA, treashold ENMO <500mg.
}
