% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicSummaryOfBouts.R
\name{basicSummaryOfBouts}
\alias{basicSummaryOfBouts}
\title{Title}
\usage{
basicSummaryOfBouts(bouts, generateCalendars, listCalendarsAndBouts,
  progreso = NULL)
}
\arguments{
\item{bouts}{}

\item{listCalendarsAndBouts}{}

\item{progreso}{}
}
\description{
Title
}
