% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeCalendars.R
\name{computeCalendars}
\alias{computeCalendars}
\title{Dado una lista de bouts, y una lista que contiene funciones de generación de calendarios,
devuelve lista de calendarios}
\usage{
computeCalendars(bouts, generateCalendars, progreso = NULL,
  idBIN = NA_character_)
}
\arguments{
\item{bouts}{}

\item{progreso}{}

\item{listFunctionalCalendars}{}
}
\description{
Dado una lista de bouts, y una lista que contiene funciones de generación de calendarios,
devuelve lista de calendarios
}
